fix(ui): Restructuring folders (AEROGEAR-8922)
