+++ ui/src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Redirect, Route } from 'react-router-dom';
 import LandingPage from './LandingPage';
-import AppDetailedView from './AppDetailedView';
+import AppViewContainer from '../containers/appView/AppViewContainer';
 
 const App = () => {
   return (
@@ -9,7 +9,7 @@ const App = () => {
       <div className="App">
         <Switch>
           <Route exact path="/" component={LandingPage} />
-          <Route path="/apps/:id" component={AppDetailedView} />
+          <Route path="/apps/:id" component={AppViewContainer} />
           {/* Default redirect */}
           <Redirect to="/" />
         </Switch>
