--- ui/src/components/AppOverview.css
@@ -1,4 +1,0 @@
-
-.app-property {
-    font-size: 1.3em;
-}
