+++ ui/src/components/appView/AppOverview.css
@@ -1,3 +1,8 @@
+
+.app-property {
+    font-size: 1.3em;
+}
+
 .app-overview {
     margin-bottom: 4rem;
 }
