+++ ui/src/components/common/TableView.js
@@ -1,21 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Table, TableHeader, TableBody } from '@patternfly/react-table';
-import { withRouter } from 'react-router-dom';
+import './TableView.css';
 
-export const AppsTable = ({ columns, rows, sortBy, onSort, onRowClick }) => (
+export const TableView = ({ columns, rows, sortBy, onSort, onRowClick }) => (
   <Table aria-label="Mobile Apps" sortBy={sortBy} onSort={onSort} cells={columns} rows={rows}>
     <TableHeader />
     <TableBody onRowClick={onRowClick}/>
   </Table>
 );
 
-AppsTable.propTypes = {
-  sortBy: PropTypes.object.isRequired,
-  onSort: PropTypes.func.isRequired,
+TableView.propTypes = {
   columns: PropTypes.array.isRequired,
   rows: PropTypes.array.isRequired,
-  onRowClick: PropTypes.func.isRequired
+  sortBy: PropTypes.object.isRequired,
+  onSort: PropTypes.func.isRequired,
+  onRowClick: PropTypes.func
 };
 
-export default withRouter(AppsTable);
+export default TableView;
